Move functionality to app
