+++ src/pages/blog.tsx
@@ -1,0 +1,14 @@
+import { AppProps } from '@/types';
+
+export default function Blog({ smallScreen, menuOpen, navOption, sectionRefs, openOrCloseMenu, updateNavOption}: AppProps) {
+
+  return (
+    <>   
+    <main className='flex flex-col'>
+    
+
+
+    </main>
+    </>
+  )
+}
