+++ src/types.ts
@@ -1,3 +1,5 @@
+import { MutableRefObject } from "react"
+
 type Indexable = {
   [index:string]: string
 }
@@ -7,4 +9,18 @@ export type Submission = Indexable & {
   email: string, 
   subject: string,
   message: string
+}
+
+export type AppProps = {
+  smallScreen: boolean,
+  menuOpen: boolean,
+  navOption: string,
+  sectionRefs: {
+    home: MutableRefObject<null>;
+    philosophy: MutableRefObject<null>;
+    services: MutableRefObject<null>;
+    contact: MutableRefObject<null>;
+  },
+  openOrCloseMenu: () => void,
+  updateNavOption: (option: string) => void,
 }
